Some refactoring of the Bessel calculation logic
